4th commit in EDA
Exploratory data analysis is the analysis of the data and brings out insights. It’s storytelling, a story that data is trying to tell. EDA is an approach to analyze the data with the help of various tools and graphical techniques like barplot, histogram, etc. There are many libraries available in python like pandas, NumPy, matplotlib, seaborn, etc. with the help of those we can do the analysis of the data and bring out helpful insights. I will be using Jupyter Notebook along with these libraries.